@@ -24,28 +24,28 @@
         "fs-extra": "^11.1.1",
         "logplease": "1.2.15",
         "minimist": "^1.2.8",
-        "semver": "^7.3.8",
+        "semver": "^7.5.4",
         "maru-ir2": "file:../maru-ir2",
-        "solc-typed-ast": "^14.0.0",
-        "web3-utils": "^1.9.0"
+        "solc-typed-ast": "^16.0.0",
+        "web3-utils": "^4.0.3"
     },
     "devDependencies": {
         "@types/fs-extra": "^11.0.1",
         "@types/minimist": "^1.2.2",
         "@types/mocha": "^10.0.1",
-        "@types/node": "^16.18.23",
-        "@types/semver": "^7.3.13",
-        "@typescript-eslint/eslint-plugin": "^5.57.0",
-        "@typescript-eslint/parser": "^5.57.0",
-        "eslint": "^8.37.0",
+        "@types/node": "^16.18.39",
+        "@types/semver": "^7.5.0",
+        "@typescript-eslint/eslint-plugin": "^6.2.0",
+        "@typescript-eslint/parser": "^6.2.0",
+        "eslint": "^8.45.0",
         "eslint-config-prettier": "^8.8.0",
-        "eslint-plugin-prettier": "^4.2.1",
-        "expect": "^29.5.0",
+        "eslint-plugin-prettier": "^5.0.0",
+        "expect": "^29.6.1",
         "mocha": "^10.2.0",
         "nyc": "^15.1.0",
-        "prettier": "^2.8.7",
+        "prettier": "^3.0.0",
         "ts-node": "^10.9.1",
-        "typescript": "^5.0.3"
+        "typescript": "^5.1.6"
     },
     "repository": "<TODO>",
     "license": "SEE LICENSE IN <LICENSE>"
